+++ src/core/lib/StorageLib.sol
@@ -13,7 +13,7 @@ struct AggregationVaultStorage {
     /// Total amount of allocation points across all strategies including the cash reserve.
     uint256 totalAllocationPoints;
     /// fee rate
-    uint256 performanceFee;
+    uint96 performanceFee;
     /// fee recipient address
     address feeRecipient;
     /// Mapping between a strategy address and it's allocation config
